refactor: split wallet verification into domain-specific hooks
- Move verifyUsernameChange to new useUserOperations hook
- Keep useWalletVerification for pure wallet operations only
